+++ com.unity.render-pipelines.core/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [TODO Next Release] - TODO Provide Date
+
+### Added
+- Added Screen Coordinates Override shader utilities.
+
 ## [14.0.0] - 2021-11-17
 
 ### Added